+++ frontend/package.json
@@ -4,6 +4,9 @@
   "private": true,
   "dependencies": {
     "@adobe/react-spectrum": "^3.16.3",
+    "@emotion/react": "^11.8.1",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.4.2",
     "@mui/material": "^5.4.3",
     "@mui/styled-engine-sc": "^5.4.2",
     "@testing-library/jest-dom": "^5.16.2",
@@ -14,7 +17,7 @@
     "react-bootstrap": "^2.1.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
+    "react-scripts": "^3.0.1",
     "react-select": "^5.2.2",
     "styled-components": "^5.3.3",
     "web-vitals": "^2.1.4"
