+++ frontend/src/App.js
@@ -6,8 +6,7 @@ import SignUp from './components/Signup/SignUp';
 import Login from './components/Login/Login';
 import NavigationBar from './components/Navbar/NavigationBar';
 import Profile from './components/Profile/Profile.js'
-
-import { getUser } from './api/apiRequest';
+import PostPage from './components/Post/PostPage';
 
 import {
   BrowserRouter as Router,
@@ -29,6 +28,7 @@ function App() {
             <Route path="/following" element={<Home/>}/>
             <Route path="/messages" element={<Home/>}/>
             <Route path="/profile" element={<Profile id={0} />}/>
+            <Route path="/post" element={<PostPage />}/>
           </Route>
           
           {/* Put pages without navbar here */}
